execute to invoke, simplified CommandMethods, invokes return actual results
