+++ builtins/src/main/java/org/jline/builtins/Commands.java
@@ -43,7 +43,6 @@ import org.jline.builtins.Completers.CompletionData;
 import org.jline.builtins.Options.HelpException;
 import org.jline.builtins.Source.StdInSource;
 import org.jline.builtins.Source.URLSource;
-import org.jline.console.ConfigurationPath;
 import org.jline.keymap.KeyMap;
 import org.jline.reader.Binding;
 import org.jline.reader.Highlighter;
