+++ console/src/main/java/org/jline/console/CommandInput.java
