+++ console/src/main/java/org/jline/console/CommandMethods.java
