+++ console/src/main/java/org/jline/console/CommandRegistry.java
