+++ console/src/main/java/org/jline/console/Printer.java
@@ -8,10 +8,13 @@
  */
 package org.jline.console;
 
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+
+import org.jline.reader.Completer;
 
 /**
  * Print object to the console.
@@ -169,12 +172,21 @@ public interface Printer {
      */
     final static String OBJECT_TO_STRING = "objectToString";
 
+    final static List<String> BOOLEAN_KEYS = Arrays.asList(ALL, ONE_ROW_TABLE, ROWNUM, SHORT_NAMES, SKIP_DEFAULT_OPTIONS
+            , STRUCT_ON_TABLE, TO_STRING);
+
     default void println(Object object) {
         println(new HashMap<>(), object);
     }
 
-    final static List<String> BOOLEAN_KEYS = Arrays.asList(ALL, ONE_ROW_TABLE, ROWNUM, SHORT_NAMES, SKIP_DEFAULT_OPTIONS
-                                                         , STRUCT_ON_TABLE, TO_STRING);
-
     void println(Map<String,Object> options, Object object);
+
+    default Exception prntCommand(CommandInput input) {
+        return null;
+    }
+
+    default List<Completer> prntCompleter(String command) {
+        return new ArrayList<>();
+    }
+
 }
