+++ console/src/main/java/org/jline/console/ScriptEngine.java
