+++ console/src/test/resources/org/jline/example/gron.nanorc
@@ -1,0 +1,12 @@
+syntax "GRON" "\.gron$"
+header "^\[$"
+
+color brightblue   "\<[-]?[0-9]*([Ee][+-]?[0-9]+)?\>"  "\<[-]?[0](\.[0-9]+)?\>"
+color yellow ""(\\.|[^"])*"|'(\\.|[^'])*'|[a-zA-Z]+[a-zA-Z0-9]*"
+color cyan  "\<null\>"
+color brightcyan "\<(true|false)\>"
+color brightyellow "\"(\\"|[^"])*\"\s*:"  "'(\'|[^'])*'\s*:" "(\[|,)\s*[a-zA-Z0-9]*\s*:"
+color white "(:|\[|,|\])"
+color magenta   "\\u[0-9a-fA-F]{4}|\\[bfnrt'"/\\]"
+color ,green "[[:space:]]+$"
+color ,red "	+ +| +	+"
